refactor: migrate validation from Zod to Yup for user forms and update related components
