+++ src/modules/admin/users/pages/CreateUserPage.tsx
@@ -8,16 +8,23 @@ import { useNotificationStore } from '@/stores';
 import CustomPageHeader from '@/components/ui/molecules/CustomPageHeader';
 import { Helmet } from 'react-helmet';
 import { FormProvider, useForm, SubmitHandler } from 'react-hook-form';
-import { z } from 'zod';
-import { zodResolver } from '@hookform/resolvers/zod';
+import * as yup from 'yup';
+import { yupResolver } from '@hookform/resolvers/yup';
 import ErrorModal2 from '@/components/ui/molecules/ErrorModal2';
 import BackPageButton from '@/components/ui/atoms/BackPageButton';
-import { UsersValidationSchema } from '../schemas/UsersAdminValidationSchema';
-import UsersFormLayout from '../components/organisms/UsersFormLayout';
+// Yup validation schema for user creation
+const UsersValidationSchema = yup.object({
+  clientId: yup.mixed().nullable().optional(),
+  username: yup.string().required('Username is required'),
+  email: yup.string().required('Email is required').email('Invalid email'),
+  fullName: yup.string().required('Full name is required').min(3, 'Full name must be at least 3 characters long').max(50, 'Full name must be at most 50 characters long'),
+  role: yup.string().oneOf(['ADMIN', 'USER'], 'Invalid role').required('Role is required'),
+});
+import UsersFormLayout from '../components/form/UsersFormLayout';
 import { formatError } from '@/lib/formatApiError.util';
 import { useRegisterUser } from '../hooks';
 
-type UserFormData = z.infer<typeof UsersValidationSchema>;
+type UserFormData = yup.InferType<typeof UsersValidationSchema>;
 
 const CreateUserPage: React.FC = () => {
   const theme = useTheme();
@@ -28,7 +35,7 @@ const CreateUserPage: React.FC = () => {
 
   const userFormMethods = useForm<UserFormData>({
     mode: 'all',
-    resolver: zodResolver(UsersValidationSchema),
+    resolver: yupResolver(UsersValidationSchema),
   });
 
   const {
